Interpolation: Add a workaround to prevent things rarely hopping around. And now I understand why.
Unfortunately, it's a fundamental flaw in the method I've chosen, since it's based on object heap memory addresses which of course can be the same for different objects across updates.
Fixing this will require either changing the SpaceObj formats (potentially save breaking) or placing hooks in the addToList() functions or similar. Hooks probably best.